fix imports related to file restructuring
