+++ syft/frameworks/torch/__init__.py
@@ -2,6 +2,5 @@ from . import tensors
 from . import federated
 from . import differential_privacy
 from . import crypto
-from . import pointers
 from .hook import hook_args
 from .hook import TorchHook
