+++ syft/frameworks/torch/tensors/interpreters/__init__.py
@@ -1,4 +1,3 @@
-from .abstract import AbstractTensor  # noqa: F401
 from .native import TorchTensor  # noqa: F401
 from .precision import FixedPrecisionTensor  # noqa: F401
 from .additive_shared import AdditiveSharingTensor  # noqa: F401
