+++ syft/frameworks/torch/tensors/interpreters/large_precision.py
@@ -3,7 +3,7 @@ import math
 import torch
 
 from syft.frameworks.torch.overload_torch import overloaded
-from syft.frameworks.torch.tensors.interpreters import AbstractTensor
+from syft.generic.tensor import AbstractTensor
 
 
 class LargePrecisionTensor(AbstractTensor):
