+++ syft/frameworks/torch/tensors/interpreters/native.py
@@ -8,8 +8,8 @@ import weakref
 
 import syft
 from syft.exceptions import InvalidTensorForRemoteGet
-from syft.frameworks.torch.tensors.interpreters import AbstractTensor
-from syft.frameworks.torch.pointers import PointerTensor
+from syft.generic.tensor import AbstractTensor
+from syft.generic.pointers import PointerTensor
 from syft.workers import BaseWorker
 from syft.frameworks.torch.tensors.interpreters.crt_precision import _moduli_for_fields
