+++ syft/frameworks/torch/tensors/interpreters/polynomial.py
@@ -1,4 +1,4 @@
-# from syft.frameworks.torch.tensors.interpreters.abstract import AbstractTensor
+# from syft.generic.tensor import AbstractTensor
 # import torch
 # import numpy as np
 # from typing import Callable, List, Union
