+++ syft/frameworks/torch/tensors/interpreters/precision.py
@@ -2,7 +2,7 @@ import torch
 
 import syft
 from syft.workers import AbstractWorker
-from syft.frameworks.torch.tensors.interpreters.abstract import AbstractTensor
+from syft.generic.tensor import AbstractTensor
 from syft.frameworks.torch.tensors.interpreters.additive_shared import AdditiveSharingTensor
 from syft.frameworks.torch.tensors.interpreters.multi_pointer import MultiPointerTensor
 from syft.frameworks.torch.overload_torch import overloaded
