+++ syft/messaging/plan.py
@@ -6,7 +6,7 @@ from typing import Union
 
 import torch
 
-from syft.frameworks.torch.tensors.interpreters.abstract import AbstractTensor
+from syft.generic.tensor import AbstractTensor
 from syft.generic import ObjectStorage
 from syft.codes import MSGTYPE
 import syft as sy
