+++ test/torch/tensors/test_native.py
@@ -2,7 +2,7 @@ import pytest
 
 import torch
 
-from syft.frameworks.torch.pointers import PointerTensor
+from syft.generic.pointers import PointerTensor
 from syft.exceptions import InvalidTensorForRemoteGet
